Refactor `add_light_gui` to header GuiSystem.h
